Changing BooksWrapper to a ts file
