--- src/Wrapper/BooksWrapper.svelte
@@ -1,10 +1,0 @@
-<script lang="ts" context="module">
-    export type Book = {
-        id: number,
-        imgUrl: string,
-        title: string,
-        authorId: number,
-        publisherId: number,
-        price: number;
-    }; 
-</script>
