+++ src/types/Book.ts
@@ -1,0 +1,8 @@
+export type Book = {
+  id: number,
+  imgUrl: string,
+  title: string,
+  authorId: number,
+  publisherId: number,
+  price: number;
+};
