+++ src/types/CartProduct.ts
@@ -1,0 +1,5 @@
+import type { Book } from "./Book";
+
+export type CartProduct = Book & {
+  amount: number;
+}
